+++ hiddenfb/test/schemas/data/wyscout/__init__.py
@@ -1,0 +1,4 @@
+from hiddenfb.test import TestUtility
+
+
+class WyscoutTestUtility(TestUtility): ...