+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Domain models for basic tracking and event data
 - Utilities to load data files
+- Utilities to validate Wyscout and Metrica event/tracking data
 
 ## [0.0.1] - 2024-01-28